Create versions of get_or_insert that don't move the key
Four new methods are added:
* get_or_insert_ref
* try_get_or_insert_ref
* get_or_insert_mut_ref
* try_get_or_insert_mut_ref

which are analog to their existing counterparts:
* get_or_insert
* try_get_or_insert
* get_or_insert_mut
* try_get_or_insert_mut

with the difference of accepting a reference to the key instead of an owned object. If the key
doesn't exist in the cache and needs to be moved, only then is it cloned using to_owned() so it can
be owned by the cache. This is useful when cloning the key is expensive.